refactor: split lib.rs into diagram, moves, raw_lines, rendering, rotation modules
