+++ src/raw_lines.rs
@@ -1,0 +1,213 @@
+use std::collections::VecDeque;
+
+use crate::rendering::Horiz;
+
+pub(crate) fn raw_lines_is_empty_above(lines: &[Vec<Horiz>], idx: usize) -> bool {
+    lines.get(idx..).unwrap_or_default().iter().all(|line| {
+        line.last()
+            .cloned()
+            .unwrap_or_default()
+            .subsequent()
+            .is_empty()
+    })
+}
+
+pub(crate) fn raw_lines_continue(lines: &mut [Vec<Horiz>]) {
+    lines
+        .iter_mut()
+        .for_each(|line| line.push(line.last().unwrap_or(&Horiz::Empty).subsequent()));
+}
+
+pub(crate) fn raw_lines_expand_above(lines: &mut [Vec<Horiz>], idx: usize) {
+    let (lower, upper) = lines.split_at_mut(idx);
+    for _ in 0..3 {
+        raw_lines_continue(lower);
+    }
+
+    let mut indexes: VecDeque<_> = upper
+        .iter_mut()
+        .map(|line| {
+            let is_empty = line
+                .last()
+                .cloned()
+                .unwrap_or_default()
+                .subsequent()
+                .is_empty();
+
+            line.push(if is_empty {
+                Horiz::Empty
+            } else {
+                Horiz::TransferUpStart
+            });
+
+            is_empty
+        })
+        .collect();
+
+    indexes.rotate_right(1);
+
+    upper
+        .iter_mut()
+        .zip(indexes.iter())
+        .for_each(|(line, is_empty)| {
+            line.push(if *is_empty {
+                Horiz::Empty
+            } else {
+                Horiz::TransferUp
+            });
+        });
+
+    indexes.rotate_right(1);
+
+    upper
+        .iter_mut()
+        .zip(indexes.iter())
+        .for_each(|(line, is_empty)| {
+            line.push(if *is_empty {
+                Horiz::Empty
+            } else {
+                Horiz::TransferUpFinish
+            });
+        });
+}
+
+pub(crate) fn raw_lines_contract_above(lines: &mut [Vec<Horiz>], idx: usize) {
+    let (lower, upper) = lines.split_at_mut(idx);
+    for _ in 0..3 {
+        raw_lines_continue(lower);
+    }
+
+    let mut indexes: VecDeque<_> = upper
+        .iter_mut()
+        .enumerate()
+        .map(|(idx, line)| {
+            let is_empty = (0..2).contains(&idx)
+                || line
+                    .last()
+                    .cloned()
+                    .unwrap_or_default()
+                    .subsequent()
+                    .is_empty();
+
+            line.push(if is_empty {
+                if idx == 0 {
+                    Horiz::ClosedAbove
+                } else if idx == 1 {
+                    Horiz::ClosedBelow
+                } else {
+                    Horiz::Empty
+                }
+            } else {
+                Horiz::TransferDownStart
+            });
+
+            is_empty
+        })
+        .collect();
+
+    indexes.rotate_left(1);
+
+    upper
+        .iter_mut()
+        .zip(indexes.iter())
+        .for_each(|(line, is_empty)| {
+            line.push(if *is_empty {
+                Horiz::Empty
+            } else {
+                Horiz::TransferDown
+            });
+        });
+
+    indexes.rotate_left(1);
+
+    upper
+        .iter_mut()
+        .zip(indexes.iter())
+        .for_each(|(line, is_empty)| {
+            line.push(if *is_empty {
+                Horiz::Empty
+            } else {
+                Horiz::TransferDownFinish
+            });
+        });
+}
+
+pub(crate) fn raw_lines_append(lines: &mut [Vec<Horiz>], element: u8, idx: usize) {
+    match element {
+        b'(' => {
+            if raw_lines_is_empty_above(&*lines, idx) {
+                raw_lines_continue(lines);
+            } else {
+                raw_lines_expand_above(lines, idx);
+            }
+            *lines[idx].last_mut().unwrap() = Horiz::OpenedAbove;
+            *lines[idx + 1].last_mut().unwrap() = Horiz::OpenedBelow;
+        }
+        b')' => {
+            let is_empty_above = raw_lines_is_empty_above(&*lines, idx + 2);
+            if is_empty_above {
+                raw_lines_continue(lines);
+                *lines[idx].last_mut().unwrap() = Horiz::ClosedAbove;
+                *lines[idx + 1].last_mut().unwrap() = Horiz::ClosedBelow;
+            } else {
+                raw_lines_contract_above(lines, idx);
+            }
+        }
+        b'\\' => {
+            raw_lines_continue(lines);
+
+            *lines[idx].last_mut().unwrap() = Horiz::CrossUpUnder;
+            *lines[idx + 1].last_mut().unwrap() = Horiz::CrossDownOver;
+        }
+        b'/' => {
+            raw_lines_continue(lines);
+
+            *lines[idx].last_mut().unwrap() = Horiz::CrossUpOver;
+            *lines[idx + 1].last_mut().unwrap() = Horiz::CrossDownUnder;
+        }
+        _ => unimplemented!(),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn snapshot_raw_lines_expand_contract() {
+        use Horiz::*;
+
+        let mut lines = vec![vec![Line], vec![Line], vec![Line], vec![Empty], vec![Empty]];
+        let original_lines = lines.clone();
+
+        raw_lines_expand_above(&mut lines, 1);
+        insta::assert_debug_snapshot!(lines);
+
+        raw_lines_contract_above(&mut lines, 1);
+        insta::assert_debug_snapshot!(lines);
+
+        raw_lines_continue(&mut lines);
+        let final_column = lines
+            .iter()
+            .map(|line| vec![line.last().cloned().unwrap()])
+            .collect::<Vec<_>>();
+        assert_eq!(final_column, original_lines);
+    }
+
+    #[test]
+    fn snapshot_raw_lines_append() {
+        let mut lines = vec![vec![]; 4];
+
+        raw_lines_append(&mut lines, b'(', 0);
+        insta::assert_debug_snapshot!(lines);
+
+        raw_lines_append(&mut lines, b'(', 1);
+        insta::assert_debug_snapshot!(lines);
+
+        raw_lines_append(&mut lines, b')', 0);
+        insta::assert_debug_snapshot!(lines);
+
+        raw_lines_append(&mut lines, b')', 0);
+        insta::assert_debug_snapshot!(lines);
+    }
+}
