+++ src/rendering.rs
@@ -1,0 +1,301 @@
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Horiz {
+    #[default]
+    Empty,
+    Line,
+    CrossDownOver,
+    CrossDownUnder,
+    CrossUpOver,
+    CrossUpUnder,
+    OpenedBelow,
+    OpenedAbove,
+    ClosedBelow,
+    ClosedAbove,
+    TransferUpStart,
+    TransferUp,
+    TransferUpFinish,
+    TransferDownStart,
+    TransferDown,
+    TransferDownFinish,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Eq)]
+pub struct VerboseLine(pub(crate) Vec<Horiz>);
+
+#[derive(Default, Debug, Clone, PartialEq, Eq)]
+pub struct VerboseDiagram(pub(crate) Vec<VerboseLine>);
+
+pub(crate) const fn display_lines(grid_borders: bool) -> usize {
+    3 + if grid_borders { 1 } else { 0 }
+}
+
+pub(crate) const DISPLAY_WITH_BORDERS_LINES: usize = display_lines(true);
+pub(crate) const DISPLAY_LINES: usize = display_lines(false);
+
+impl Horiz {
+    #[rustfmt::skip]
+    pub const fn display(&self) -> [&'static str; DISPLAY_LINES] {
+        use Horiz::*;
+
+        match self {
+            Empty => [
+                r#"   "#,
+                r#"   "#,
+                r#"   "#,
+            ],
+            Line => [
+                r#"___"#,
+                r#"   "#,
+                r#"   "#,
+            ],
+            CrossDownOver => [
+                r#"   "#,
+                r#"\ /"#,
+                r#" \ "#,
+            ],
+            CrossDownUnder => [
+                r#"   "#,
+                r#"\ /"#,
+                r#" / "#,
+            ],
+            CrossUpOver | CrossUpUnder => [
+                r#"/ \"#,
+                r#"   "#,
+                r#"   "#,
+            ],
+            OpenedBelow => [
+                r#"   "#,
+                r#"  /"#,
+                r#" ( "#,
+            ],
+            OpenedAbove => [
+                r#"  \"#,
+                r#"   "#,
+                r#"   "#,
+            ],
+            ClosedBelow => [
+                r#"   "#,
+                r#"\  "#,
+                r#" ) "#,
+            ],
+            ClosedAbove => [
+                r#"/  "#,
+                r#"   "#,
+                r#"   "#,
+            ],
+            TransferUpStart => [
+                r#"__/"#,
+                r#"   "#,
+                r#"   "#,
+            ],
+            TransferUp => [
+                r#"  /"#,
+                r#" / "#,
+                r#"/  "#,
+            ],
+            TransferUpFinish => [
+                r#"  _"#,
+                r#" / "#,
+                r#"/  "#,
+            ],
+            TransferDownStart => [
+                r#"_  "#,
+                r#" \ "#,
+                r#"  \"#,
+            ],
+            TransferDown => [
+                r#"\  "#,
+                r#" \ "#,
+                r#"  \"#,
+            ],
+            TransferDownFinish => [
+                r#"\__"#,
+                r#"   "#,
+                r#"   "#,
+            ],
+        }
+    }
+
+    #[rustfmt::skip]
+    pub const fn display_with_borders(&self) -> [&'static str; DISPLAY_WITH_BORDERS_LINES] {
+        use Horiz::*;
+
+        match self {
+            Empty => [
+                r#"+---"#,
+                r#"|   "#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+            Line => [
+                r#"+---"#,
+                r#"|___"#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+            CrossDownOver => [
+                r#"+---"#,
+                r#"|   "#,
+                r#"|\ /"#,
+                r#"| \ "#,
+            ],
+            CrossDownUnder => [
+                r#"+---"#,
+                r#"|   "#,
+                r#"|\ /"#,
+                r#"| / "#,
+            ],
+            CrossUpOver | CrossUpUnder => [
+                r#"+---"#,
+                r#"|/ \"#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+            OpenedBelow => [
+                r#"+---"#,
+                r#"|   "#,
+                r#"|  /"#,
+                r#"| ( "#,
+            ],
+            OpenedAbove => [
+                r#"+---"#,
+                r#"|  \"#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+            ClosedBelow => [
+                r#"+---"#,
+                r#"|   "#,
+                r#"|\  "#,
+                r#"| ) "#,
+            ],
+            ClosedAbove => [
+                r#"+---"#,
+                r#"|/  "#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+            TransferUpStart => [
+                r#"+---"#,
+                r#"|__/"#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+            TransferUp => [
+                r#"+---"#,
+                r#"|  /"#,
+                r#"| / "#,
+                r#"|/  "#,
+            ],
+            TransferUpFinish => [
+                r#"+---"#,
+                r#"|  _"#,
+                r#"| / "#,
+                r#"|/  "#,
+            ],
+            TransferDownStart => [
+                r#"+---"#,
+                r#"|_  "#,
+                r#"| \ "#,
+                r#"|  \"#,
+            ],
+            TransferDown => [
+                r#"+---"#,
+                r#"|\  "#,
+                r#"| \ "#,
+                r#"|  \"#,
+            ],
+            TransferDownFinish => [
+                r#"+---"#,
+                r#"|\__"#,
+                r#"|   "#,
+                r#"|   "#,
+            ],
+        }
+    }
+
+    #[rustfmt::skip]
+    pub const fn subsequent(&self) -> Self {
+        use Horiz::*;
+
+        match self {
+            | Empty
+            | ClosedBelow
+            | ClosedAbove
+            | TransferUpStart
+            | TransferUp
+            | TransferDownStart
+            | TransferDown
+            => Empty,
+
+            | Line
+            | CrossDownOver
+            | CrossDownUnder
+            | CrossUpOver
+            | CrossUpUnder
+            | OpenedBelow
+            | OpenedAbove
+            | TransferUpFinish
+            | TransferDownFinish
+            => Line,
+        }
+    }
+
+    pub const fn is_empty(&self) -> bool {
+        matches!(self, Horiz::Empty)
+    }
+}
+
+impl VerboseLine {
+    pub fn display<const GRID_BORDERS: bool>(&self) -> impl 'static + Iterator<Item = String> {
+        let horiz_len: usize = if GRID_BORDERS {
+            Horiz::Empty.display_with_borders()[0].len()
+        } else {
+            Horiz::Empty.display()[0].len()
+        };
+
+        let mut l0 = " ".repeat(self.0.len() * horiz_len) + "\n";
+        let mut l1 = l0.clone();
+        let mut l2 = l0.clone();
+        let mut l3 = if GRID_BORDERS {
+            l0.clone()
+        } else {
+            String::new()
+        };
+
+        for (idx, horiz) in self.0.iter().enumerate() {
+            let [h0, h1, h2, h3] = if GRID_BORDERS {
+                horiz.display_with_borders()
+            } else {
+                let [h0, h1, h2] = horiz.display();
+                [h0, h1, h2, ""]
+            };
+            let range = (idx * horiz_len)..((idx + 1) * horiz_len);
+
+            l0.replace_range(range.clone(), h0);
+            l1.replace_range(range.clone(), h1);
+            l2.replace_range(range.clone(), h2);
+            if GRID_BORDERS {
+                l3.replace_range(range, h3);
+            }
+        }
+
+        [l0, l1, l2]
+            .into_iter()
+            .chain(std::iter::once(l3).filter(|_| GRID_BORDERS))
+    }
+}
+
+impl VerboseDiagram {
+    pub fn display<'a, const GRID_BORDERS: bool>(&'a self) -> impl 'a + Iterator<Item = String> {
+        let (last_idx, inner) = match self.0.len().checked_sub(1) {
+            Some(idx) => (idx, self.0.as_slice()),
+            None => (0, &[][..]),
+        };
+
+        inner.iter().rev().enumerate().flat_map(move |(idx, line)| {
+            line.display::<GRID_BORDERS>()
+                .take(display_lines(GRID_BORDERS) - if idx == last_idx { 2 } else { 0 })
+        })
+    }
+}
